Revert app.tsx
