+++ client/src/App.tsx
@@ -11,20 +11,6 @@ import SampleContainer from "./components/examples/SampleContainer";
 import SignInModal from "./pages/SignInModal";
 import SignUpModal from "./pages/SignUpModal";
 
-import FormItem from './components/molecules/FormItem'
-import Modal from './components/organisms/Modal'
-
-import { FunctionComponent } from 'react'
-
-interface Props {
-  isDisabled: boolean;
-}
-const Bob: FunctionComponent<Props> = (props: Props) => {
-  return (
-    <input type="text" disabled={props.isDisabled}/>
-  )
-}
-
 function App(): JSX.Element {
   return (
     <span>
@@ -40,16 +26,6 @@ function App(): JSX.Element {
           <Route path='/test'>
             <RequestGroupDonorView requestGroupId="603d9b41eb57fc06447b8a23" />
           </Route>
-          <Route path='/formitem'>
-              <FormItem 
-                formItemName="Group Name" 
-                errorString="Please enter a group name" 
-                isDisabled={false} 
-                inputComponent={<Bob isDisabled={false}/>}
-                tooltipText="Groups describe overall cateogry of item such as stroller crib bed"
-              />
-
-              </Route>
           <Route path='/'><DonorHomepage /></Route>
         </Switch>
       </Router>
