+++ client/src/components/molecules/style/FormItem.scss
@@ -9,6 +9,7 @@
         display: flex;
         flex-direction: row;
         justify-content: space-between;
+        width: 300px;
     }
     &-bottom {
         text-align: left;
